bazel fix. avx2 opt defined only if ICC

+++ cpp/oneapi/dal/algo/jaccard/detail/service.hpp
@@ -29,10 +29,10 @@ ONEDAL_FORCEINLINE std::int32_t max(const std::int32_t &a, const std::int32_t &b
 }
 
 ONEDAL_FORCEINLINE std::int64_t compute_number_elements_in_block(
-    const std::int32_t &row_range_begin,
-    const std::int32_t &row_range_end,
-    const std::int32_t &column_range_begin,
-    const std::int32_t &column_range_end) {
+    const std::int64_t &row_range_begin,
+    const std::int64_t &row_range_end,
+    const std::int64_t &column_range_begin,
+    const std::int64_t &column_range_end) {
     ONEDAL_ASSERT(row_range_end >= row_range_begin, "Negative interval found");
     const std::int64_t row_count = row_range_end - row_range_begin;
     ONEDAL_ASSERT(column_range_end >= column_range_begin, "Negative interval found");
